Fix directory name
