+++ internal/views/topic.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"time"
 
-	"godiscourse/internal/model"
+	"godiscourse/internal/models"
 )
 
 // TopicView is the response body of topic
@@ -24,7 +24,7 @@ type TopicView struct {
 	Category      CategoryView `json:"category"`
 }
 
-func buildTopic(topic *model.Topic) TopicView {
+func buildTopic(topic *models.Topic) TopicView {
 	view := TopicView{
 		Type:          "topic",
 		TopicID:       topic.TopicID,
@@ -44,12 +44,12 @@ func buildTopic(topic *model.Topic) TopicView {
 }
 
 // RenderTopic response a topic
-func RenderTopic(w http.ResponseWriter, r *http.Request, topic *model.Topic) {
+func RenderTopic(w http.ResponseWriter, r *http.Request, topic *models.Topic) {
 	RenderResponse(w, r, buildTopic(topic))
 }
 
 // RenderTopics response a bundle of topics
-func RenderTopics(w http.ResponseWriter, r *http.Request, topics []*model.Topic) {
+func RenderTopics(w http.ResponseWriter, r *http.Request, topics []*models.Topic) {
 	topicViews := make([]TopicView, len(topics))
 	for i, t := range topics {
 		topicViews[i] = buildTopic(t)
